split out common funcs and testcases
